% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pn.R
\name{Pn}
\alias{Pn}
\title{Pn Function}
\usage{
Pn(lambda, mu, n)
}
\arguments{
\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{n}{number of items}
}
\value{
Probability
}
\description{
Probability of having n items in the system
}
\examples{
Pn(lambda=5, mu=6, n=3)
}
