% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PW.R
\name{PW}
\alias{PW}
\title{PW Function}
\usage{
PW(lambda, mu, x)
}
\arguments{
\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{x}{time}
}
\value{
Probability
}
\description{
Returns the Probability that the cycle time is larger than x
}
\examples{
PW(lambda=5, mu=6, x=2)
}
