% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWq.R
\name{PWq}
\alias{PWq}
\title{PWq Function}
\usage{
PWq(lambda, mu, x)
}
\arguments{
\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{x}{time}
}
\value{
Probability
}
\description{
Returns the Probability that the waiting time is larger than x
}
\examples{
PWq(lambda=5, mu=6, x=2)
}
